odh-sre-agent/pyproject.toml
[project]
name = "odh-sre-agent"
version = "0.1.0"
description = "Multi-agent SRE system for ROSA OpenShift + Open Data Hub using Claude Agent SDK"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.40",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "models", "tools"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["agents", "models", "tools", "config"]
